% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds4ci_hello.R
\name{ds4ci_hello}
\alias{ds4ci_hello}
\title{Say hello to World or stated person/group.}
\usage{
ds4ci_hello(to_who = "World")
}
\arguments{
\item{to_who}{string}
}
\value{
"Hello [World | <to_who>] from DS4CI"
}
\description{
Say hello to World or stated person/group.
}
\examples{
ds4ci_hello()
ds4ci_hello("Fred")
}

